backend/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
regex==2024.11.6
pillow==11.1.0
django-cors-headers==4.6.0
drf-yasg==1.21.8